requirements.txt
argh==0.26.2
Babel==2.5.3
cov-core==1.15.0
coverage==4.5.1
Django==2.0.3
django-model-utils==3.1.1
django-money==0.12.3
django-phonenumber-field==2.0.0
djangorestframework==3.7.7
flake8==2.2.2
ghp-import==0.5.5
Jinja2==2.10
Markdown==2.4.1
MarkupSafe==1.0
mccabe==0.6.1
mkdocs==0.11.1
pathtools==0.1.2
pep8==1.7.1
phonenumberslite==8.9.1
pkg-resources==0.0.0
py==1.5.2
py-moneyed==0.7.0
pyflakes==1.6.0
pytest==2.5.2
pytest-cov==1.6
pytest-django==2.6
pytz==2018.3
PyYAML==3.12
watchdog==0.8.3
